Move last few files
