+++ trview.app/Menus/AlternateGroupToggler.cpp
