+++ trview.app/Menus/AlternateGroupToggler.h
