+++ trview.app/Settings/UserSettings.cpp
