+++ trview.app/Settings/UserSettings.h
