+++ trview.app/Windows/RouteWindow.cpp
@@ -1,6 +1,6 @@
 #include "RouteWindow.h"
-#include <trview.app/Waypoint.h>
-#include <trview.app/Route.h>
+#include <trview.app/Routing/Waypoint.h>
+#include <trview.app/Routing/Route.h>
 #include <trview.ui/GroupBox.h>
 #include <trview.ui/TextArea.h>
 #include <trview.ui/Button.h>
