+++ trview.app/trview.app.vcxproj
@@ -19,7 +19,6 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="AlternateGroupToggler.cpp" />
     <ClCompile Include="Camera\Camera.cpp" />
     <ClCompile Include="Camera\CameraMode.cpp" />
     <ClCompile Include="Camera\FreeCamera.cpp" />
@@ -38,12 +37,15 @@
     <ClCompile Include="Graphics\ITextureStorage.cpp" />
     <ClCompile Include="Graphics\SectorHighlight.cpp" />
     <ClCompile Include="Graphics\SelectionRenderer.cpp" />
+    <ClCompile Include="Menus\AlternateGroupToggler.cpp" />
     <ClCompile Include="Menus\DirectoryListing.cpp" />
     <ClCompile Include="Menus\FileDropper.cpp" />
     <ClCompile Include="Menus\LevelSwitcher.cpp" />
     <ClCompile Include="Menus\RecentFiles.cpp" />
     <ClCompile Include="Menus\ViewMenu.cpp" />
-    <ClCompile Include="Route.cpp" />
+    <ClCompile Include="Routing\Route.cpp" />
+    <ClCompile Include="Routing\Waypoint.cpp" />
+    <ClCompile Include="Settings\UserSettings.cpp" />
     <ClCompile Include="Tools\Compass.cpp" />
     <ClCompile Include="Tools\Measure.cpp" />
     <ClCompile Include="Tools\Toolbar.cpp" />
@@ -57,8 +59,6 @@
     <ClCompile Include="UI\RoomNavigator.cpp" />
     <ClCompile Include="UI\SettingsWindow.cpp" />
     <ClCompile Include="UI\ViewerUI.cpp" />
-    <ClCompile Include="UserSettings.cpp" />
-    <ClCompile Include="Waypoint.cpp" />
     <ClCompile Include="Windows\CollapsiblePanel.cpp" />
     <ClCompile Include="Windows\ItemsWindow.cpp" />
     <ClCompile Include="Windows\ItemsWindowManager.cpp" />
@@ -69,7 +69,6 @@
     <ClCompile Include="Windows\WindowResizer.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="AlternateGroupToggler.h" />
     <ClInclude Include="Camera\Camera.h" />
     <ClInclude Include="Camera\CameraMode.h" />
     <ClInclude Include="Camera\FreeCamera.h" />
@@ -93,12 +92,15 @@
     <ClInclude Include="Graphics\ITextureStorage.h" />
     <ClInclude Include="Graphics\SectorHighlight.h" />
     <ClInclude Include="Graphics\SelectionRenderer.h" />
+    <ClInclude Include="Menus\AlternateGroupToggler.h" />
     <ClInclude Include="Menus\DirectoryListing.h" />
     <ClInclude Include="Menus\FileDropper.h" />
     <ClInclude Include="Menus\LevelSwitcher.h" />
     <ClInclude Include="Menus\RecentFiles.h" />
     <ClInclude Include="Menus\ViewMenu.h" />
-    <ClInclude Include="Route.h" />
+    <ClInclude Include="Routing\Route.h" />
+    <ClInclude Include="Routing\Waypoint.h" />
+    <ClInclude Include="Settings\UserSettings.h" />
     <ClInclude Include="Tools\Compass.h" />
     <ClInclude Include="Tools\Measure.h" />
     <ClInclude Include="Tools\Toolbar.h" />
@@ -112,8 +114,6 @@
     <ClInclude Include="UI\RoomNavigator.h" />
     <ClInclude Include="UI\SettingsWindow.h" />
     <ClInclude Include="UI\ViewerUI.h" />
-    <ClInclude Include="UserSettings.h" />
-    <ClInclude Include="Waypoint.h" />
     <ClInclude Include="Windows\CollapsiblePanel.h" />
     <ClInclude Include="Windows\ItemsWindow.h" />
     <ClInclude Include="Windows\ItemsWindowManager.h" />
